Include invisible `https://` in value
Mastodon's `Account::Field` class does some weird stuff to determine
whether or not a particular field's value is "verifiable." Among those
checks is a test that the element is an anchor and that it's `href`
attribute matches the node's text. Weird.

Mastodon proper meets this requirement as it publishes field values that
look like:

```
<a href="…"><span class="invisible">https://</span>…</a>
```

So weird. This should hopefully help with this profile's URL(s) being
shown as verified.